refactor: update integer literal to store int64

+++ ast/integer_literal.go
@@ -4,7 +4,7 @@ import "taulang/token"
 
 type IntegerLiteral struct {
 	Token token.Token
-	Value int
+	Value int64
 }
 
 func (i *IntegerLiteral) expressionNode() {}
